repin ECommons.IPC 帶進跨外掛 IPC 宣告修正,並收掉 Gearsetter 的 SourceInventorySlot 型別窄化
舊 pin 71759390 停在 2026-08-26 歷史壓縮之前,推進到 7e40218 之後帶進四筆修正:
  - Questionable 的 AddQuestPriority 與 ClearQuestPriority 原本互相對調(呼叫必擲 IpcLengthMismatchError)
  - AutoRetainer.PluginState.EnqueueHET 由 Action<bool,bool> 改為與提供端一致的 Action<Action>
  - Lifestream.GetResidentialTerritory 改用逐值核對過的鏡像列舉
  - AutoDuty 自己的 SetConfig 訂閱端簽章修正(舊 pin 也漏掉了這一筆)

同一波收掉 Gearsetter 的型別窄化:提供端回傳的 tuple 第三元是 int?,
ECommons.IPC 訂閱端與這裡的包裝方法都宣告成 byte?。值不超過 255 時 Newtonsoft
靜默轉換成功,超過 255 才擲 JsonSerializationException 並一路落到 IpcTypeMismatchError,
而 SafeWrapper.IPCException 攔不住它。目前格位索引踩不到,但屬於「只在資料變大那一次炸」的形狀。

IPCSubscriber.cs 的包裝方法與 AutoEquipHelper.cs 的欄位/解構同步改成 int?;
下游只有 (int)sourceInventorySlot 的轉型,行為完全不變。

祖先閘門(兩步都過才 repin):
  - 舊 pin 71759390 是 tag pre-squash-20260826 (82d6b110) 的祖先,遠端可達
  - squash 根 99b6c9b 的 tree 與該 tag 的 tree 位元組同一 (f5fab62fbe67750687ce1745412210e7f0ea2a8d)

+++ AutoDuty/Helpers/AutoEquipHelper.cs
@@ -112,7 +112,7 @@ namespace AutoDuty.Helpers
             }
         }
 
-        private List<(uint ItemId, InventoryType? SourceInventory, byte? SourceInventorySlot, RaptureGearsetModule.GearsetItemIndex TargetSlot)>? _gearset           = null;
+        private List<(uint ItemId, InventoryType? SourceInventory, int? SourceInventorySlot, RaptureGearsetModule.GearsetItemIndex TargetSlot)>? _gearset           = null;
         private int                                                                                                                               _index             = 0;
 
         private void AutoEquipGearSetterUpdate(IFramework framework)
@@ -144,7 +144,7 @@ namespace AutoDuty.Helpers
             }
             else if (this._gearset != null && this._index < this._gearset.Count)
             {
-                (uint itemId, InventoryType? inventoryType, byte? sourceInventorySlot, RaptureGearsetModule.GearsetItemIndex targetSlot) = this._gearset[this._index];
+                (uint itemId, InventoryType? inventoryType, int? sourceInventorySlot, RaptureGearsetModule.GearsetItemIndex targetSlot) = this._gearset[this._index];
                 DebugLog($"Equip item {itemId} in {targetSlot} from {inventoryType} (slot {sourceInventorySlot})");
 
                 if (inventoryType != null && sourceInventorySlot != null)
